Add support for tools finding tools from the path
